WIP - move a bucketload of logic down into Items component
